+++ backend-api/middleware/studentAuthMiddleware.js
